Enable use of prebuilt Qt in installer builds
When we create Qt online installers and IFW installers, we build Qt
every time. As the Qt version changes rarely, more optimized way is to
build Qt once and use that. This change enables downloading prebuilt
compressed Qt binaries and use that for building installers. Especially
this speeds up jenkins jobs where we cannot use --incremental flag.

Task-number: QTQAINFRA-2074

+++ packaging-tools/bld_ifw_tools.py
@@ -151,6 +151,8 @@ class IfwOptions:
                  qt_installer_framework_qmake_args,
                  openssl_dir,
                  product_key_checker_pri,
+                 qt_binaries_static,
+                 qt_binaries_dynamic,
                  incremental_build = False):
         self.incremental_mode                           = incremental_build
         self.qt_source_dir                              = os.path.join(ROOT_DIR, 'qt-src')
@@ -166,6 +168,12 @@ class IfwOptions:
         self.qt_installer_framework_branch_pretty       = qt_installer_framework_branch.replace("/", "_")
         self.qt_installer_framework_qmake_args          = qt_installer_framework_qmake_args
         self.openssl_dir                                = openssl_dir
+        self.qt_binaries_static                         = qt_binaries_static
+        if self.qt_binaries_static:
+            self.qt_binaries_static_saveas              = os.path.join(ROOT_DIR, os.path.basename(self.qt_binaries_static))
+        self.qt_binaries_dynamic                        = qt_binaries_dynamic
+        if self.qt_binaries_dynamic:
+            self.qt_binaries_dynamic_saveas             = os.path.join(ROOT_DIR, os.path.basename(self.qt_binaries_dynamic))
         self.qt_build_modules                           = " module-qtbase module-qtdeclarative module-qttools module-qttranslations"
         self.qt_build_modules_docs                      = " module-qttools"
         if bldinstallercommon.is_win_platform():
@@ -255,6 +263,8 @@ class IfwOptions:
         print('installer_framework_target_dir:          {0}'.format(self.installer_framework_target_dir))
         print('installer_framework_payload_arch:        {0}'.format(self.installer_framework_payload_arch))
         print('product_key_checker:                     {0}'.format(self.product_key_checker_pri))
+        print('qt_binaries_static:                      {0}'.format(self.qt_binaries_static))
+        print('qt_binaries_dynamic:                     {0}'.format(self.qt_binaries_dynamic))
         print('-----------------------------------------')
 
 
@@ -269,23 +279,30 @@ def build_ifw(options, create_installer=False, build_ifw_examples=False):
     clean_build_environment(options)
     #checkout sources
     prepare_installer_framework(options)
-    # copy qt sources
+    # copy qt sources. Also when fetching qt binaries as there are stuff the qt binaries need
     prepare_qt_sources(options)
-    # build qt
-    build_qt(options, options.qt_build_dir, options.qt_configure_options, options.qt_build_modules)
+    # extract qt binaries if qt binaries found, else build qt
+    if options.qt_binaries_static:
+        prepare_compressed_package(options.qt_binaries_static, options.qt_binaries_static_saveas, options.qt_build_dir)
+    else:
+        build_qt(options, options.qt_build_dir, options.qt_configure_options, options.qt_build_modules)
     # build installer framework
     build_installer_framework(options)
     if build_ifw_examples:
         build_installer_framework_examples(options)
     # steps when creating ifw installer
     if create_installer:
-        configure_options = get_dynamic_qt_configure_options() + '-prefix ' + options.qt_build_dir_dynamic + os.sep + 'qtbase'
-        # Although we have a shadow build qt sources are still taminated. Unpack sources again.
-        if os.path.exists(options.qt_source_dir):
-            bldinstallercommon.remove_tree(options.qt_source_dir)
-        # copy qt sources
-        prepare_qt_sources(options)
-        build_qt(options, options.qt_build_dir_dynamic, configure_options, options.qt_build_modules_docs)
+        if options.qt_binaries_dynamic:
+            prepare_compressed_package(options.qt_binaries_dynamic, options.qt_binaries_dynamic_saveas, options.qt_build_dir_dynamic)
+            if bldinstallercommon.is_win_platform():
+                patch_win32_mkspecs(os.path.join(options.qt_build_dir_dynamic, "qtbase", "mkspecs"))
+        else:
+            configure_options = get_dynamic_qt_configure_options() + '-prefix ' + options.qt_build_dir_dynamic + os.sep + 'qtbase'
+            # Although we have a shadow build qt sources are still taminated. Unpack sources again.
+            if os.path.exists(options.qt_source_dir):
+                bldinstallercommon.remove_tree(options.qt_source_dir)
+            prepare_qt_sources(options)
+            build_qt(options, options.qt_build_dir_dynamic, configure_options, options.qt_build_modules_docs)
         build_ifw_docs(options)
         create_installer_package(options)
     #archive
@@ -302,7 +319,7 @@ def prepare_qt_sources(options):
         return
     print('--------------------------------------------------------------------')
     print('Prepare Qt src package: {0}'.format(options.qt_source_package_uri))
-    prepare_src_package(options.qt_source_package_uri, options.qt_source_package_uri_saveas, options.qt_source_dir)
+    prepare_compressed_package(options.qt_source_package_uri, options.qt_source_package_uri_saveas, options.qt_source_dir)
 
     if bldinstallercommon.is_win_platform():
         patch_win32_mkspecs(os.path.join(options.qt_source_dir, "qtbase", "mkspecs"))
@@ -311,8 +328,8 @@ def prepare_qt_sources(options):
 ###############################
 # function
 ###############################
-def prepare_src_package(src_pkg_uri, src_pkg_saveas, destination_dir):
-    print('Fetching Src package from: {0}'.format(src_pkg_uri))
+def prepare_compressed_package(src_pkg_uri, src_pkg_saveas, destination_dir):
+    print('Fetching package from: {0}'.format(src_pkg_uri))
     if not os.path.isfile(src_pkg_saveas):
         if not bldinstallercommon.is_content_url_valid(src_pkg_uri):
             print('*** Src package uri is invalid! Abort!')
@@ -376,7 +393,7 @@ def prepare_installer_framework(options):
         bldinstallercommon.clone_repository(options.qt_installer_framework_uri, options.qt_installer_framework_branch, options.installer_framework_source_dir)
     else:
         # fetch src package
-        prepare_src_package(options.qt_installer_framework_uri, options.qt_installer_framework_uri_saveas, options.installer_framework_source_dir)
+        prepare_compressed_package(options.qt_installer_framework_uri, options.qt_installer_framework_uri_saveas, options.installer_framework_source_dir)
 
 
 ###############################
@@ -658,6 +675,8 @@ def setup_argument_parser():
     parser.add_argument('--create_installer', help="Build the ifw and wrap it in the installer", action='store_true', required=False, default=False)
     parser.add_argument('--incremental', help="Build the ifw in incremental mode", action='store_true', required=False, default=False)
     parser.add_argument('--build_ifw_examples', help="Build the IFW examples", action='store_true', required=False, default=False)
+    parser.add_argument('--qt_binaries_static', help="Use prebuilt Qt package instead of building from scratch", required=False)
+    parser.add_argument('--qt_binaries_dynamic', help="Use prebuilt Qt package instead of building from scratch", required=False)
     return parser
 
 
@@ -687,6 +706,8 @@ if __name__ == "__main__":
                          ifw_qmake_args,
                          CARGS.openssl_dir,
                          CARGS.product_key_checker_pri,
+                         CARGS.qt_binaries_static,
+                         CARGS.qt_binaries_dynamic,
                          CARGS.incremental
                         )
     # build ifw tools
